+++ api/src/main/java/io/strimzi/api/kafka/model/listener/arraylistener/GenericKafkaListenerConfiguration.java
@@ -205,8 +205,7 @@ public class GenericKafkaListenerConfiguration implements Serializable, UnknownP
             "`SingleStack` is for a single IP family. " +
             "`PreferDualStack` is for two IP families on dual-stack configured clusters or a single IP family on single-stack clusters. " +
             "`RequireDualStack` fails unless there are two IP families on dual-stack configured clusters. " +
-            "If unspecified, Kubernetes will choose the default value based on the service type. " +
-            "Available on Kubernetes 1.20 and newer.")
+            "If unspecified, Kubernetes will choose the default value based on the service type.")
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @PresentInVersions("v1beta2+")
     public IpFamilyPolicy getIpFamilyPolicy() {
@@ -219,8 +218,7 @@ public class GenericKafkaListenerConfiguration implements Serializable, UnknownP
 
     @Description("Specifies the IP Families used by the service. " +
             "Available options are `IPv4` and `IPv6. " +
-            "If unspecified, Kubernetes will choose the default value based on the `ipFamilyPolicy` setting. " +
-            "Available on Kubernetes 1.20 and newer.")
+            "If unspecified, Kubernetes will choose the default value based on the `ipFamilyPolicy` setting.")
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @PresentInVersions("v1beta2+")
     public List<IpFamily> getIpFamilies() {