+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaConnectorIT.java
@@ -121,7 +121,7 @@ public class KafkaConnectorIT {
     public void testConnectorNotUpdatedWhenConfigUnchanged(VertxTestContext context) {
         KafkaConnectApiImpl connectClient = new KafkaConnectApiImpl(vertx);
 
-        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_20);
+        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_21);
 
         String namespace = "ns";
         String connectorName = "my-connector";
@@ -190,7 +190,7 @@ public class KafkaConnectorIT {
     public void testConnectorResourceNotReadyWhenConnectorFailed(VertxTestContext context) {
         KafkaConnectApiImpl connectClient = new KafkaConnectApiImpl(vertx);
 
-        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_20);
+        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_21);
 
         String namespace = "ns";
         String connectorName = "my-connector-2";
@@ -239,7 +239,7 @@ public class KafkaConnectorIT {
     public void testConnectorResourceNotReadyWhenTaskFailed(VertxTestContext context) {
         KafkaConnectApiImpl connectClient = new KafkaConnectApiImpl(vertx);
 
-        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_20);
+        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_21);
 
         String namespace = "ns";
         String connectorName = "my-connector-3";
@@ -299,7 +299,7 @@ public class KafkaConnectorIT {
     public void testConnectorIsAutoRestarted(VertxTestContext context) {
         KafkaConnectApiImpl connectClient = new KafkaConnectApiImpl(vertx);
 
-        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_20);
+        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_21);
 
         String namespace = "ns";
         String connectorName = "my-connector-4";
@@ -348,7 +348,7 @@ public class KafkaConnectorIT {
     public void testTaskIsAutoRestarted(VertxTestContext context) {
         KafkaConnectApiImpl connectClient = new KafkaConnectApiImpl(vertx);
 
-        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_20);
+        PlatformFeaturesAvailability pfa = new PlatformFeaturesAvailability(false, KubernetesVersion.V1_21);
 
         String namespace = "ns";
         String connectorName = "my-connector-5";